+++ CHANGES.rst
@@ -6,6 +6,15 @@ Version 3.2.0
 Unreleased
 
 
+Version 3.1.1
+-------------
+
+Released 2022-03-25
+
+-   The template filename on Windows uses the primary path separator.
+    :issue:`1637`
+
+
 Version 3.1.0
 -------------